+++ features/readwrite-splitting/distsql/parser/src/main/antlr4/imports/readwrite-splitting/Keyword.g4
@@ -134,23 +134,3 @@ WEIGHT
 NOT
     : N O T
     ;
-
-QUERY
-    : Q U E R Y
-    ;
-
-LOAD
-    : L O A D
-    ;
-
-BALANCE
-    : B A L A N C E
-    ;
-
-ALGORITHM
-    : A L G O R I T H M
-    ;
-
-PLUGINS
-    : P L U G I N S
-    ;
