+++ features/readwrite-splitting/distsql/parser/src/main/antlr4/imports/readwrite-splitting/RALStatement.g4
@@ -26,7 +26,3 @@ alterReadwriteSplittingStorageUnitStatus
 showStatusFromReadwriteSplittingRules
     : SHOW STATUS FROM READWRITE_SPLITTING (RULES | RULE ruleName) (FROM databaseName)?
     ;
-
-showReadQueryLoadBalanceAlgorithmPlugins
-    : SHOW READ QUERY LOAD BALANCE ALGORITHM PLUGINS
-    ;
