+++ infra/distsql-handler/src/main/resources/META-INF/services/org.apache.shardingsphere.distsql.handler.executor.ral.plugin.PluginTypeAndClassMapper
@@ -16,3 +16,4 @@
 #
 
 org.apache.shardingsphere.distsql.handler.executor.ral.plugin.type.KeyGenerateAlgorithmTypeAndClassMapper
+org.apache.shardingsphere.distsql.handler.executor.ral.plugin.type.LoadBalanceAlgorithmTypeAndClassMapper
