+++ parser/distsql/engine/src/main/antlr4/org/apache/shardingsphere/distsql/parser/autogen/KernelDistSQLStatement.g4
@@ -50,5 +50,6 @@ execute
     | unlockCluster
     | showPluginImplementations
     | showKeyGenerateAlgorithmPlugins
+    | showLoadBalanceAlgorithmPlugins
     ) SEMI_? EOF
     ;
